# go: Add omitempty support and struct as map