--- a/Go/sereal/decode.go
+++ b/Go/sereal/decode.go
@@ -985,15 +985,15 @@ func (d *Decoder) decodeHashViaReflection(by []byte, idx int, ln int, ptr reflec
 				return 0, err
 			}
 
-			var fld int
+			var fld tag
 			var found bool
 
 			if tags == nil {
 				// do nothing
 			} else if fld, found = tags[string(key)]; found {
-				idx, err = d.decodeViaReflection(by, idx, ptr.Field(fld))
+				idx, err = d.decodeViaReflection(by, idx, ptr.Field(fld.id))
 			} else if fld, found = tags[strings.Title(string(key))]; found {
-				idx, err = d.decodeViaReflection(by, idx, ptr.Field(fld))
+				idx, err = d.decodeViaReflection(by, idx, ptr.Field(fld.id))
 			}
 
 			if !found {