--- a/Go/sereal/sereal_test.go
+++ b/Go/sereal/sereal_test.go
@@ -776,6 +776,16 @@ func TestStructAsMap(t *testing.T) {
 			AOmitTags{},
 			map[string]interface{}{},
 		},
+		{
+			"Half",
+			AOmitTags{Phone: "12345"},
+			map[string]interface{}{"phone": "12345"},
+		},
+		{
+			"Full",
+			AOmitTags{Name: "mr foo", Phone: "12345"},
+			map[string]interface{}{"name": "mr foo", "phone": "12345"},
+		},
 	}
 
 	for _, compat := range []bool{false, true} {
@@ -820,6 +830,9 @@ func TestStructAsMap(t *testing.T) {
 				}
 				if !reflect.DeepEqual(routStruct.Elem().Interface(), routMap.Elem().Interface()) {
 					t.Errorf("roundtrip mismatch for %s: got: %#v expected: %#v\n", v.what, routStruct.Elem().Interface(), routMap.Elem().Interface())
+				}
+				if !reflect.DeepEqual(routStruct.Elem().Interface(), rinputMap.Interface()) {
+					t.Errorf("source mismatch for %s: got: %#v expected: %#v\n", v.what, routStruct.Elem().Interface(), rinputMap.Interface())
 				}
 			})
 		}