Those features are meant to reduce the resulting binary size of Marshal result.

The `omitempty` part is nearly the same as `encoding/json` support.