@sapk 2020-11-17
Gentle ping @dgryski if you have time to review this suggestion. 
You seems to be the last contributor on the go part.