@dgryski 2020-11-17
I'm only tangentially involved in this project at the moment.  Probably best to get a developer still at Booking.com to review this.