@sapk 2024-08-13
I wanted to compare benchmark with master they are kind of broken because we have a struct with un-exported fields

What I can provide is in this PR, activating StructAsMap provide this improvement:
```
name                                    old time/op    new time/op    delta
EncodeComplexDataWithHeader-8             10.3µs ± 3%     9.4µs ± 2%   -8.71%  (p=0.000 n=10+9)
EncodeAndSnappyComplexDataWithHeader-8    11.1µs ± 4%    10.1µs ± 4%   -9.26%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8      56.1µs ± 1%    53.0µs ± 2%   -5.60%  (p=0.000 n=10+8)

name                                    old bytes      new bytes      delta
EncodeComplexDataWithHeader-8               528B ± 1%      503B ± 1%   -4.64%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8      532B ± 4%      506B ± 2%   -4.89%  (p=0.000 n=9+10)
EncodeAndZlibComplexDataWithHeader-8        454B ± 1%      444B ± 1%   -2.14%  (p=0.000 n=10+10)

name                                    old alloc/op   new alloc/op   delta
EncodeComplexDataWithHeader-8             4.29kB ± 0%    3.81kB ± 0%  -11.19%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8    4.91kB ± 0%    4.40kB ± 0%  -10.35%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8      6.72kB ± 3%    5.96kB ± 4%  -11.26%  (p=0.000 n=9+9)

name                                    old allocs/op  new allocs/op  delta
EncodeComplexDataWithHeader-8               95.0 ± 0%      94.0 ± 0%   -1.05%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8      96.0 ± 0%      95.0 ± 0%   -1.04%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8         101 ± 0%       100 ± 0%   -0.99%  (p=0.000 n=10+10)
```

With fixing benchmark in master if we compare without StructAsMap (to see impact of omitempty only):
```
name                                    old time/op    new time/op    delta
EncodeComplexDataWithHeader-8             10.3µs ±39%    10.3µs ± 3%     ~     (p=0.165 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8    9.80µs ± 7%   11.12µs ± 4%  +13.43%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8      53.9µs ± 1%    56.1µs ± 1%   +4.11%  (p=0.000 n=8+10)

name                                    old bytes      new bytes      delta
EncodeComplexDataWithHeader-8               542B ± 0%      528B ± 1%   -2.68%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8      547B ± 1%      532B ± 4%   -2.72%  (p=0.000 n=9+9)
EncodeAndZlibComplexDataWithHeader-8        463B ± 1%      454B ± 1%   -2.05%  (p=0.000 n=10+10)

name                                    old alloc/op   new alloc/op   delta
EncodeComplexDataWithHeader-8             4.34kB ± 0%    4.29kB ± 0%   -1.11%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8    4.98kB ± 0%    4.91kB ± 0%   -1.35%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8      6.61kB ± 2%    6.72kB ± 3%   +1.68%  (p=0.021 n=8+9)

name                                    old allocs/op  new allocs/op  delta
EncodeComplexDataWithHeader-8               97.0 ± 0%      95.0 ± 0%   -2.06%  (p=0.000 n=10+10)
EncodeAndSnappyComplexDataWithHeader-8      98.0 ± 0%      96.0 ± 0%   -2.04%  (p=0.000 n=10+10)
EncodeAndZlibComplexDataWithHeader-8         103 ± 0%       101 ± 0%   -1.94%  (p=0.000 n=10+10)
```
It seems to take a bit more time (but there is a lot of variation on this measure) and the other-side it take less allocation.